backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# AWS
boto3==1.34.34

# OCR & Image Processing
#pytesseract==0.3.10
#Pillow>=10.0.0
#opencv-python==4.9.0.80
#PyMuPDF==1.23.8
#rapidfuzz==3.14.3

pytesseract==0.3.10
Pillow>=10.0.0
PyMuPDF==1.23.8
rapidfuzz==3.14.3

# ML/AI
#--extra-index-url https://download.pytorch.org/whl/cpu
#torch==2.10.0+cpu
#torchvision==0.25.0+cpu

# Face Recognition (V2)
#insightface>=0.7.3
#onnxruntime>=1.16.0
#scikit-learn>=1.3.0
#opencv-python-headless>=4.9.0,<4.10  # insightface dependency, must be <4.10 for numpy<2 compat

insightface>=0.7.3
onnxruntime>=1.16.0
scikit-learn>=1.3.0
opencv-python-headless>=4.9.0,<4.10

# HTTP Client
httpx==0.27.0

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Excel
openpyxl==3.1.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Geocoding
geopy>=2.4.1
googlemaps>=4.10.0

# NumPy <2.0 required for torch/insightface compatibility
numpy>=1.24.0,<2.0